third_party/blink/web_tests/platform/generic/http/tests/xmlhttprequest/workers/post-formdata-expected.txt

[Worker] Test verifies that FormData is sent correctly when using XMLHttpRequest asynchronously.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Starting worker: ./resources/post-formdata-worker.js
PASS [Worker] echoResult is "string=string value"
PASS [Worker] echoResult is "bareBlob=blob:application/octet-stream:blob-value"
PASS [Worker] echoResult is "mimeBlob=blob:text/html:blob-value"
PASS [Worker] echoResult is "namedBlob=blob-file.txt:application/octet-stream:blob-value"
PASS [Worker] echoResult is "bareFile=file-name.txt:application/octet-stream:file-value"
PASS [Worker] echoResult is "mimeFile=file-name.html:text/html:file-value"
PASS [Worker] echoResult is "renamedFile=file-name-override.html:text/html:file-value"
PASS successfullyParsed is true

TEST COMPLETE
